# src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserProfileResponse": {
				"properties": {
					"techStack": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"coverImgUrl": {
						"type": "string",
						"nullable": true
					},
					"profileImgUrl": {
						"type": "string",
						"nullable": true
					},
					"portfolioLink": {
						"type": "string",
						"nullable": true
					},
					"bio": {
						"type": "string",
						"nullable": true
					},
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string",
						"nullable": true
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"techStack",
					"coverImgUrl",
					"profileImgUrl",
					"portfolioLink",
					"bio",
					"email",
					"username",
					"name",
					"id"
				],
				"type": "object"
			},
			"ApiResponse_UserProfileResponse_": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserProfileResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"statusCode",
					"status",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"errors": {
						"items": {},
						"type": "array",
						"default": []
					}
				},
				"required": [
					"name",
					"message",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserEditInput": {
				"properties": {
					"portfolioLink": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"ApiResponse____": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {},
						"type": "object",
						"nullable": true
					}
				},
				"required": [
					"statusCode",
					"status",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserSummaryResponse": {
				"properties": {
					"techStack": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"bio": {
						"type": "string",
						"nullable": true
					},
					"coverImgUrl": {
						"type": "string",
						"nullable": true
					},
					"profileImgUrl": {
						"type": "string",
						"nullable": true
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"username": {
						"type": "string",
						"nullable": true
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"techStack",
					"bio",
					"coverImgUrl",
					"profileImgUrl",
					"email",
					"name",
					"username",
					"id"
				],
				"type": "object"
			},
			"ApiResponse_UserSummaryResponse-Array_": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/UserSummaryResponse"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"statusCode",
					"status",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FollowResponse": {
				"properties": {
					"isFollowing": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"isFollowing",
					"message"
				],
				"type": "object"
			},
			"ApiResponse_FollowResponse_": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/FollowResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"statusCode",
					"status",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserFollowerInput": {
				"properties": {
					"followerId": {
						"type": "string"
					},
					"followingId": {
						"type": "string"
					}
				},
				"required": [
					"followerId",
					"followingId"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "UNNERA API",
		"version": "1.0.0",
		"description": "API Docs for UNERRA",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "utkarsh yadav"
		}
	},
	"paths": {
		"/user/update": {
			"patch": {
				"operationId": "UpdateUserProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserProfileResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					},
					"409": {
						"description": "Email or username already exists",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Update user profile",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserEditInput"
							}
						}
					}
				}
			}
		},
		"/user/{userId}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserProfileResponse_"
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Get user by ID",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse____"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Delete user",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserSummaryResponse-Array_"
								}
							}
						}
					}
				},
				"description": "Get users with pagination",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/user/follow": {
			"post": {
				"operationId": "FollowUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FollowResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Follow or unfollow a user",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserFollowerInput"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}